send email notifications when changing affiliations, split up test files
