+++ api_tests/institutions/views/test_institution_relationship_nodes.py
@@ -1,8 +1,8 @@
 import pytest
+from unittest import mock
 
 from api.base.settings.defaults import API_BASE
 from osf_tests.factories import (
-    WithdrawnRegistrationFactory,
     RegistrationFactory,
     InstitutionFactory,
     AuthUserFactory,
@@ -51,7 +51,7 @@ class TestInstitutionRelationshipNodes:
         return node
 
     @pytest.fixture()
-    def node_two(self, user):
+    def node_without_institution(self, user):
         return NodeFactory(creator=user)
 
     @pytest.fixture()
@@ -62,6 +62,14 @@ class TestInstitutionRelationshipNodes:
         return node_public
 
     @pytest.fixture()
+    def admin(self, institution, node_public):
+        user = AuthUserFactory()
+        user.add_or_update_affiliated_institution(institution)
+        node_public.add_contributor(user, permissions=permissions.ADMIN)
+        user.save()
+        return user
+
+    @pytest.fixture()
     def node_private(self, user, institution):
         node_private = NodeFactory()
         node_private.affiliated_institutions.add(institution)
@@ -70,14 +78,10 @@ class TestInstitutionRelationshipNodes:
 
     @pytest.fixture()
     def url_institution_nodes(self, institution):
-        return '/{}institutions/{}/relationships/nodes/'.format(
-            API_BASE, institution._id)
+        return f'/{API_BASE}institutions/{institution._id}/relationships/nodes/'
 
-    def test_auth_get_nodes(
-            self, app, user, node_one, node_public,
-            node_private, url_institution_nodes
-    ):
-        # test_get_nodes_no_auth
+    def test_get_nodes_no_auth(self, app, user, node_one, node_public, node_private, url_institution_nodes):
+
         res = app.get(url_institution_nodes)
 
         assert res.status_code == 200
@@ -86,7 +90,8 @@ class TestInstitutionRelationshipNodes:
         assert node_public._id in node_ids
         assert node_private._id not in node_ids
 
-        # test_get_nodes_with_auth
+    def test_get_nodes_with_auth(self, app, user, node_one, node_public, node_private, url_institution_nodes):
+
         res = app.get(url_institution_nodes, auth=user.auth)
 
         assert res.status_code == 200
@@ -95,9 +100,7 @@ class TestInstitutionRelationshipNodes:
         assert node_public._id in node_ids
         assert node_private._id not in node_ids
 
-    def test_node_or_type_does_not_exist(
-            self, app, user, node_two, url_institution_nodes):
-        # test_node_does_not_exist
+    def test_node_does_not_exist(self, app, user, node_without_institution, url_institution_nodes):
         res = app.post_json_api(
             url_institution_nodes,
             make_payload('notIdatAll'),
@@ -106,31 +109,33 @@ class TestInstitutionRelationshipNodes:
 
         assert res.status_code == 404
 
+    def test_node_type_does_not_exist(self, app, user, node_without_institution, url_institution_nodes):
+
         # test_node_type_does_not_exist
         res = app.post_json_api(
             url_institution_nodes,
-            {'data': [{'type': 'dugtrio', 'id': node_two._id}]},
+            {'data': [{'type': 'dugtrio', 'id': node_without_institution._id}]},
             expect_errors=True, auth=user.auth
         )
 
         assert res.status_code == 409
 
     def test_user_with_nodes_and_permissions(
-            self, user, app, node_two,
+            self, user, app, node_without_institution,
             url_institution_nodes, institution
     ):
         res = app.post_json_api(
             url_institution_nodes,
-            make_payload(node_two._id),
+            make_payload(node_without_institution._id),
             auth=user.auth
         )
 
         assert res.status_code == 201
         node_ids = [node_['id'] for node_ in res.json['data']]
-        assert node_two._id in node_ids
+        assert node_without_institution._id in node_ids
 
-        node_two.reload()
-        assert institution in node_two.affiliated_institutions.all()
+        node_without_institution.reload()
+        assert institution in node_without_institution.affiliated_institutions.all()
 
     def test_user_does_not_have_node(
             self, app, node,
@@ -148,18 +153,18 @@ class TestInstitutionRelationshipNodes:
         assert institution not in node.affiliated_institutions.all()
 
     def test_user_is_admin(
-            self, app, node_two,
+            self, app, node_without_institution,
             url_institution_nodes, user,
             institution
     ):
         res = app.post_json_api(
             url_institution_nodes,
-            make_payload(node_two._id),
+            make_payload(node_without_institution._id),
             auth=user.auth
         )
         assert res.status_code == 201
-        node_two.reload()
-        assert institution in node_two.affiliated_institutions.all()
+        node_without_institution.reload()
+        assert institution in node_without_institution.affiliated_institutions.all()
 
     def test_user_is_read_write(
             self, app, node,
@@ -215,36 +220,36 @@ class TestInstitutionRelationshipNodes:
         assert institution not in node.affiliated_institutions.all()
 
     def test_add_some_with_permissions_others_without(
-            self, user, node, node_two, app, url_institution_nodes, institution):
+            self, user, node, node_without_institution, app, url_institution_nodes, institution):
         res = app.post_json_api(
             url_institution_nodes,
-            make_payload(node_two._id, node._id),
+            make_payload(node_without_institution._id, node._id),
             expect_errors=True, auth=user.auth
         )
 
         assert res.status_code == 403
-        node_two.reload()
+        node_without_institution.reload()
         node.reload()
-        assert institution not in node_two.affiliated_institutions.all()
+        assert institution not in node_without_institution.affiliated_institutions.all()
         assert institution not in node.affiliated_institutions.all()
 
     def test_add_some_existant_others_not(
             self, institution, node_one, user,
-            node_two, app, url_institution_nodes
+            node_without_institution, app, url_institution_nodes
     ):
         assert institution in node_one.affiliated_institutions.all()
 
         res = app.post_json_api(
             url_institution_nodes,
-            make_payload(node_two._id, node_one._id),
+            make_payload(node_without_institution._id, node_one._id),
             auth=user.auth
         )
 
         assert res.status_code == 201
         node_one.reload()
-        node_two.reload()
+        node_without_institution.reload()
         assert institution in node_one.affiliated_institutions.all()
-        assert institution in node_two.affiliated_institutions.all()
+        assert institution in node_without_institution.affiliated_institutions.all()
 
     def test_only_add_existent_with_mixed_permissions(
             self, institution, node_one, node_public,
@@ -266,18 +271,18 @@ class TestInstitutionRelationshipNodes:
         assert institution in node_public.affiliated_institutions.all()
 
     def test_only_add_existent_with_permissions(
-            self, user, node_one, node_two,
+            self, user, node_one, node_without_institution,
             institution, app,
             url_institution_nodes
     ):
-        node_two.affiliated_institutions.add(institution)
-        node_two.save()
+        node_without_institution.affiliated_institutions.add(institution)
+        node_without_institution.save()
         assert institution in node_one.affiliated_institutions.all()
-        assert institution in node_two.affiliated_institutions.all()
+        assert institution in node_without_institution.affiliated_institutions.all()
 
         res = app.post_json_api(
             url_institution_nodes,
-            make_payload(node_two._id, node_one._id),
+            make_payload(node_without_institution._id, node_one._id),
             auth=user.auth
         )
 
@@ -386,288 +391,53 @@ class TestInstitutionRelationshipNodes:
 
         assert res.status_code == 404
 
+    def test_email_sent_on_affiliation_addition(self, app, user, institution, node_without_institution,
+                                                url_institution_nodes):
+        node_without_institution.add_contributor(user, permissions='admin')
+        current_institution = InstitutionFactory()
+        node_without_institution.affiliated_institutions.add(current_institution)
 
-@pytest.mark.django_db
-class TestInstitutionRelationshipRegistrations:
+        with mock.patch('osf.models.mixins.mails.send_mail') as mocked_send_mail:
+            res = app.post_json_api(
+                url_institution_nodes,
+                {
+                    'data': [
+                        {
+                            'type': 'nodes', 'id': node_without_institution._id
+                        }
+                    ]
+                },
+                auth=user.auth
+            )
 
-    @pytest.fixture()
-    def institution(self):
-        return InstitutionFactory()
+            assert res.status_code == 201
+            mocked_send_mail.assert_called_once()
+            call_args = mocked_send_mail.call_args[1]
+            assert institution in call_args['institution_added']
+            assert set([current_institution, institution]) == set(call_args['current_affiliations'])
 
-    @pytest.fixture()
-    def admin(self, institution):
-        user = AuthUserFactory()
-        user.add_or_update_affiliated_institution(institution)
-        user.save()
-        return user
+    def test_email_sent_on_affiliation_removal(self, app, admin, institution, node_public, url_institution_nodes):
+        current_institution = InstitutionFactory()
+        node_public.affiliated_institutions.add(current_institution)
 
-    @pytest.fixture()
-    def user(self, institution):
-        return AuthUserFactory()
+        with mock.patch('osf.models.mixins.mails.send_mail') as mocked_send_mail:
+            res = app.delete_json_api(
+                url_institution_nodes,
+                {
+                    'data': [
+                        {
+                            'type': 'nodes', 'id': node_public._id
+                        }
+                    ]
+                },
+                auth=admin.auth
+            )
 
-    @pytest.fixture()
-    def affiliated_user(self, institution):
-        user = AuthUserFactory()
-        user.add_or_update_affiliated_institution(institution)
-        user.save()
-        return user
+            # Assert response is successful
+            assert res.status_code == 204
 
-    @pytest.fixture()
-    def registration_no_owner(self):
-        return RegistrationFactory(is_public=True)
-
-    @pytest.fixture()
-    def registration_no_affiliation(self, admin):
-        return RegistrationFactory(creator=admin)
-
-    @pytest.fixture()
-    def registration_pending(self, admin, institution):
-        registration = RegistrationFactory(creator=admin)
-        registration.affiliated_institutions.add(institution)
-        registration.save()
-        return registration
-
-    @pytest.fixture()
-    def registration_public(self, admin, institution):
-        registration = RegistrationFactory(creator=admin, is_public=True)
-        registration.affiliated_institutions.add(institution)
-        registration.save()
-        return registration
-
-    @pytest.fixture()
-    def url_institution_registrations(self, institution):
-        return '/{}institutions/{}/relationships/registrations/'.format(
-            API_BASE, institution._id)
-
-    def test_auth_get_registrations(
-            self, app, admin, registration_no_owner,
-            registration_no_affiliation,
-            registration_pending, registration_public,
-            url_institution_registrations
-    ):
-
-        # test getting registrations without auth (for complete registrations)
-        res = app.get(url_institution_registrations)
-        assert res.status_code == 200
-        registration_ids = [reg['id'] for reg in res.json['data']]
-        assert registration_no_affiliation._id not in registration_ids
-        assert registration_pending._id not in registration_ids
-        assert registration_no_owner._id not in registration_ids
-        assert registration_public._id in registration_ids
-
-        # Withdraw a registration, make sure it still shows up
-        WithdrawnRegistrationFactory(
-            registration=registration_public, user=admin)
-
-        # test getting registrations with auth (for embargoed and pending)
-        res = app.get(url_institution_registrations, auth=admin.auth)
-        assert res.status_code == 200
-        registration_ids = [reg['id'] for reg in res.json['data']]
-        assert registration_no_affiliation._id not in registration_ids
-        assert registration_pending._id in registration_ids
-        assert registration_public._id in registration_ids
-        assert registration_no_owner._id not in registration_ids
-
-    def test_add_incorrect_permissions(
-            self, app, admin, user, affiliated_user,
-            registration_no_affiliation,
-            url_institution_registrations,
-            institution
-    ):
-        # No authentication
-        res = app.post_json_api(
-            url_institution_registrations,
-            make_registration_payload(registration_no_affiliation._id),
-            expect_errors=True,
-        )
-        assert res.status_code == 401
-
-        # User has no permission
-        res = app.post_json_api(
-            url_institution_registrations,
-            make_registration_payload(registration_no_affiliation._id),
-            expect_errors=True, auth=AuthUserFactory().auth
-        )
-        assert res.status_code == 403
-
-        # User has read permission
-        registration_no_affiliation.add_contributor(
-            affiliated_user, permissions=permissions.READ)
-        registration_no_affiliation.save()
-
-        res = app.post_json_api(
-            url_institution_registrations,
-            make_registration_payload(registration_no_affiliation._id),
-            auth=user.auth, expect_errors=True
-        )
-        assert res.status_code == 403
-
-        # User is admin but not affiliated
-        registration = RegistrationFactory(creator=user)
-        res = app.post_json_api(
-            url_institution_registrations,
-            make_registration_payload(registration._id),
-            expect_errors=True, auth=user.auth
-        )
-        assert res.status_code == 403
-        registration.reload()
-        assert institution not in registration.affiliated_institutions.all()
-
-        # Registration does not exist
-        res = app.post_json_api(
-            url_institution_registrations,
-            make_registration_payload('notIdatAll'),
-            expect_errors=True, auth=admin.auth
-        )
-        assert res.status_code == 404
-
-        # Attempt to use endpoint on Node
-        res = app.post_json_api(
-            url_institution_registrations,
-            {'data': [{'type': 'nodes', 'id': NodeFactory(creator=admin)._id}]},
-            expect_errors=True, auth=admin.auth
-        )
-        assert res.status_code == 409
-
-        registration_no_affiliation.reload()
-        assert institution not in registration_no_affiliation.affiliated_institutions.all()
-
-    def test_add_some_with_permissions_others_without(
-            self, admin, app,
-            registration_no_affiliation,
-            registration_no_owner,
-            url_institution_registrations,
-            institution
-    ):
-        res = app.post_json_api(
-            url_institution_registrations,
-            make_registration_payload(
-                registration_no_owner._id,
-                registration_no_affiliation._id),
-            expect_errors=True,
-            auth=admin.auth)
-
-        assert res.status_code == 403
-        registration_no_owner.reload()
-        registration_no_affiliation.reload()
-        assert institution not in registration_no_owner.affiliated_institutions.all()
-        assert institution not in registration_no_affiliation.affiliated_institutions.all()
-
-    def test_add_user_is_admin(
-            self, admin, app,
-            registration_no_affiliation,
-            url_institution_registrations,
-            institution
-    ):
-        res = app.post_json_api(
-            url_institution_registrations,
-            make_registration_payload(registration_no_affiliation._id),
-            auth=admin.auth
-        )
-
-        assert res.status_code == 201
-        registration_no_affiliation.reload()
-        assert institution in registration_no_affiliation.affiliated_institutions.all()
-
-    def test_add_withdrawn_registration(
-            self, app, url_institution_registrations,
-            admin, registration_no_affiliation,
-            institution
-    ):
-        WithdrawnRegistrationFactory(
-            registration=registration_no_affiliation, user=admin)
-
-        res = app.post_json_api(
-            url_institution_registrations,
-            make_registration_payload(registration_no_affiliation._id),
-            auth=admin.auth
-        )
-
-        assert res.status_code == 201
-        registration_no_affiliation.reload()
-        assert institution in registration_no_affiliation.affiliated_institutions.all()
-
-    def test_add_user_is_read_write(
-            self, app, affiliated_user,
-            registration_no_affiliation,
-            url_institution_registrations,
-            institution
-    ):
-        registration_no_affiliation.add_contributor(affiliated_user)
-        registration_no_affiliation.save()
-        res = app.post_json_api(
-            url_institution_registrations,
-            make_registration_payload(registration_no_affiliation._id),
-            auth=affiliated_user.auth
-        )
-
-        assert res.status_code == 201
-        registration_no_affiliation.reload()
-        assert institution in registration_no_affiliation.affiliated_institutions.all()
-
-    def test_add_already_added(
-            self, admin, app,
-            registration_pending,
-            url_institution_registrations,
-            institution
-    ):
-        res = app.post_json_api(
-            url_institution_registrations,
-            make_registration_payload(registration_pending._id),
-            auth=admin.auth
-        )
-
-        assert res.status_code == 204
-        registration_pending.reload()
-        assert institution in registration_pending.affiliated_institutions.all()
-
-    def test_delete_user_is_admin(
-            self, app,
-            url_institution_registrations,
-            registration_pending,
-            admin, institution
-    ):
-        res = app.delete_json_api(
-            url_institution_registrations,
-            make_registration_payload(registration_pending._id),
-            auth=admin.auth
-        )
-        registration_pending.reload()
-        assert res.status_code == 204
-        assert institution not in registration_pending.affiliated_institutions.all()
-
-    def test_delete_user_is_read_write(
-            self, app, affiliated_user,
-            registration_pending,
-            url_institution_registrations,
-            institution
-    ):
-        registration_pending.add_contributor(affiliated_user)
-        registration_pending.save()
-
-        res = app.delete_json_api(
-            url_institution_registrations,
-            make_registration_payload(registration_pending._id),
-            auth=affiliated_user.auth
-        )
-        registration_pending.reload()
-
-        assert res.status_code == 204
-        assert institution not in registration_pending.affiliated_institutions.all()
-
-    def test_delete_user_is_admin_but_not_affiliated_with_inst(
-            self, user, institution, app, url_institution_registrations):
-        registration = RegistrationFactory(creator=user)
-        registration.affiliated_institutions.add(institution)
-        registration.save()
-        assert institution in registration.affiliated_institutions.all()
-
-        res = app.delete_json_api(
-            url_institution_registrations,
-            make_registration_payload(registration._id),
-            auth=user.auth
-        )
-
-        assert res.status_code == 204
-        registration.reload()
-        assert institution not in registration.affiliated_institutions.all()
+            call_args = mocked_send_mail.call_args[1]
+            assert call_args['current_affiliations'] == [current_institution]
+            assert call_args['institution_removed'] == [institution]
+            assert call_args['user'] == admin
+            assert node_public == call_args['node']
